DHLGW-202: handling no service selected

+++ Model/Rest/CheckoutDataManagement.php
@@ -117,6 +117,9 @@ class CheckoutDataManagement implements CheckoutDataManagementInterface
      */
     public function setServiceSelection(string $quoteId, array $serviceSelection)
     {
+        if (empty($serviceSelection)) {
+            return;
+        }
         $addressId = (string)$this->shippingAdressManagement->get((int)$quoteId)->getId();
         $this->quoteServiceSelectionRepository->deleteByQuoteAddressId($addressId);
